refactor(ui): move all styled-components to separate style.tsx files
