+++ StudyPlannerUI/src/components/Modal/style.tsx
@@ -8,8 +8,8 @@
  * the full text of the GNU General Public License.
  */
 
-import styled from 'styled-components';
 import { Tokens } from 'style/tokens';
+import styled from 'styled-components';
 
 export const ModalDialog = styled.dialog`
   background: ${Tokens.neutral20};
@@ -55,4 +55,13 @@ export const CloseButton = styled.button`
     }
     scale: 0.95;
   }
+`;
+export const ToastContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 5px 20px;
+  border-radius: 4px;
+  border: 1px solid #000;
+  background-color ${({ theme }) => theme.primaryContainer};
 `;
