+++ StudyPlannerUI/src/components/Tooltip/index.tsx
@@ -9,49 +9,8 @@
  */
 
 import React, { useRef, useState } from 'react';
-import styled from 'styled-components';
 
-const TooltipContainer = styled.div<{ top: string; left: string }>`
-  visibility: hidden;
-  position: absolute;
-  top: 100%;
-  font-size: 0.8rem;
-  margin-top: 10px;
-  // top: ${props => props.top};
-  // left: ${props => props.left};
-  background-color: ${({ theme }) => theme.secondary};
-  color: ${({ theme }) => theme.onSecondary};
-  padding: 10px;
-  border-radius: 5px;
-  z-index: 100;
-  white-space: nowrap;
-  transition: opacity 0.3s;
-  opacity: 0;
-  &:after {
-    content: '';
-    position: absolute;
-    top: -25%;
-    left: ${props => props.left};
-    // left: 50%;
-    transform: rotate(180deg);
-    // transform: translateX(-50%);
-    border-width: 5px;
-    border-style: solid;
-    border-color: ${({ theme }) => theme.secondary} transparent transparent transparent;
-  }
-`;
-
-const TooltipWrapper = styled.div`
-  display: inline-flex;
-  position: relative;
-  visibility: visible;
-  opacity: 1;
-
-  &:hover ${TooltipContainer} {
-    visibility: visible;
-    opacity: 1;
-  }
-`;
+import { TooltipContainer, TooltipWrapper } from './style';
 
 interface TooltipProps {
   children: React.ReactNode;
